Tidy up some comments.
Add a (void)expr to #define hash, to avoid a compiler warning about
unused argument keysize (which is sometimes a variable only used in
this context).

Remove some pointless inline specs. Inline is nice when building the
library itself, but is of no use to the library's clients. Good
compilers (gcc -O3) will inline stuff anyway (though not across
libraries). In any case, this stuff can wait for performance
measurement.